+++ src/lucky_migrator/create_table_statement.cr
@@ -108,7 +108,7 @@ class LuckyMigrator::CreateTableStatement
     {% foreign_key_name = underscored_class + "_id" %}
     %table_name = {{ references }} || LuckyInflector::Inflector.pluralize({{ underscored_class }})
 
-    add_column :{{ foreign_key_name }}, Int64, {{ optional }}, reference: %table_name, on_delete: {{ on_delete }}
+    add_column :{{ foreign_key_name }}, Int32, {{ optional }}, reference: %table_name, on_delete: {{ on_delete }}
     add_index :{{ foreign_key_name }}
   end
